Revert "Move Chain to its own directory (#208)"
This reverts commit 0cb9d9811d8c76d29e666d69e2faa83db00a3304.
